\name{exams-package}
\alias{exams-package}
\alias{exams}
\docType{package}
\title{
\packageTitle{exams}
}
\description{
\packageDescription{exams}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{exams}
\packageIndices{exams}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{exams}

Maintainer: \packageMaintainer{exams}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
