\name{examstat-package}
\alias{examstat-package}
\alias{examstat}
\docType{package}
\title{
\packageTitle{examstat}
}
\description{
\packageDescription{examstat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{examstat}
\packageIndices{examstat}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{examstat}

Maintainer: \packageMaintainer{examstat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
