% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNotes.R
\name{getNotes}
\alias{getNotes}
\title{Get total score (on 100) and grades}
\usage{
getNotes(dt, coefNote, firstQ = "Q1", lastQ = "C12")
}
\arguments{
\item{dt}{The dataframe in which the grades are recorded by student and by question}

\item{coefNote}{The multiplying factor to convert the raw total exam score to the score on 100}

\item{firstQ}{Name of the column corresponding to the first question}

\item{lastQ}{Name of the column corresponding to the last question}
}
\value{
The dataframe with the total score on 100 and the corresponding grade
}
\description{
Compute the total score on 100 and add grades according to the UQTR, Psychology Dpt rules
}
\examples{
notes <- getNotes(dt, 2, "Q1", "C16");
}
