% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histTot.R
\name{histTot}
\alias{histTot}
\title{Plot the exam scores as an histogram}
\usage{
histTot(dt, Total = "Total", xMin = 35, xMax = 50)
}
\arguments{
\item{dt}{The dataframe in with the total scores of the exam by student}

\item{Total}{Name of the column corresponding to the total score of the exam}

\item{xMin}{minimum value to display on the x-axis [optional]}

\item{xMax}{minimum value to display on the x-axis [optional]}
}
\value{
The histogram
}
\description{
Plot the exam scores as an histogram
}
\examples{
notes <- histTot(dt, Total="Total");
}
