% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcentQ.R
\name{pcentQ}
\alias{pcentQ}
\title{Compute the percentage of success for each question}
\usage{
pcentQ(ptMax, dt)
}
\arguments{
\item{ptMax}{A dataframe with the maximum score possible for each question}

\item{dt}{The dataframe in with the raw scores by student and by question}
}
\value{
The barplot
}
\description{
Compute the percentage of success for each question
}
\examples{
notes <- pcentQ(ptMax=pointsMax, dt=questions);
}
