% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotByQuest.R
\name{plotByQuest}
\alias{plotByQuest}
\title{Plot each question (raw score)}
\usage{
plotByQuest(dt)
}
\arguments{
\item{dt}{The dataframe in with the total scores of the exam by student}

\item{mean}{Name of the column with the mean per question}

\item{quest}{Name of the column with the different questions}
}
\value{
The barplot
}
\description{
This function runs on the dataframe from statQuestion()
}
\examples{
notes <- plotByQuest(dt);
}
