% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotByQuestPcent.R
\name{plotByQuestPcent}
\alias{plotByQuestPcent}
\title{Plot the percentage of success for each question, by question type}
\usage{
plotByQuestPcent(dt, QType = FALSE, threshold = 70)
}
\arguments{
\item{dt}{The dataframe in with the percentage of success of each question is reported}

\item{QType}{A logical value to display per question type [optional]}

\item{threshold}{A numeric value to display a horizontal line as threshold of success}
}
\value{
The barplot
}
\description{
This function runs on the output of the pcentQ() function
}
\examples{
plotByQuestPcent(dt);
}
