% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statExam.R
\name{statExam}
\alias{statExam}
\title{Get the descriptive statistics of a given exam}
\usage{
statExam(dt, note = "Note", exam = NULL)
}
\arguments{
\item{dt}{The dataframe in which the grades are recorded by student and by question}

\item{note}{Name of the column where to find the total score of the exams}

\item{exam}{Name of the column defining different exams (as factor) [optional]}
}
\value{
A dataframe with the descriptive statistics by exam
}
\description{
Compute the mean, standard deviation, min and max score of a given exam
}
\examples{
notes <- statExam(dt, note="Note");
notes <- statExam(dt, note="Note", exam="ExamType");
}
