% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statQuestions.R
\name{statQuestions}
\alias{statQuestions}
\title{Get the descriptive statistics by question}
\usage{
statQuestions(dt, firstQ = "Q1", lastQ = "C16")
}
\arguments{
\item{dt}{The dataframe in which the grades are recorded by student and by question}

\item{firstQ}{Name of the column corresponding to the first question}

\item{lastQ}{Name of the column corresponding to the last question}
}
\value{
A dataframe with the descriptive statistics by question
}
\description{
Compute the mean, standard deviation, min and max score for each question of a given exam
}
\examples{
notes <- statQuestions(dt, firstQ="Q1", lastQ="C16");
}
